+++ src/main/java/cc/cassian/raspberry/registry/RaspberryTags.java
@@ -4,6 +4,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
 import net.minecraftforge.registries.ForgeRegistries;
 
 import static cc.cassian.raspberry.RaspberryMod.MOD_ID;
@@ -18,6 +19,11 @@ public class RaspberryTags {
     public static final TagKey<Item> WORM_SEEKER_ITEMS = createItemTag( "worm_seeker_items");
     public static final TagKey<EntityType<?>> WORM_SEEKERS = createEntityTypeTag( "worm_seekers");
 
+    public static final TagKey<Block> SHEARS_SHOULD_MINE = createBlockTag("shears_should_mine");
+
+    public static TagKey<Block> createBlockTag(String id) {
+        return TagKey.create(ForgeRegistries.BLOCKS.getRegistryKey(), new ResourceLocation(MOD_ID, id));
+    }
 
     public static TagKey<Item> createItemTag(String id) {
         return TagKey.create(ForgeRegistries.ITEMS.getRegistryKey(), new ResourceLocation(MOD_ID, id));